+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -67,6 +67,7 @@
 #import "SentryMigrateSessionInit.h"
 #import "SentryNSError.h"
 #import "SentryNSURLRequest.h"
+#import "SentryNetworkTracker.h"
 #import "SentryOutOfMemoryLogic.h"
 #import "SentryOutOfMemoryTracker.h"
 #import "SentryPerformanceTracker.h"
@@ -101,3 +102,4 @@
 #import "SentryUIViewControllerSanitizer.h"
 #import "SentryUserFeedback.h"
 #import "TestSentryCrashAdapter.h"
+#import "URLSessionTaskMock.h"